Add EQ http endpoint
If the EQ is not on then pressing SET reads from the eq.txt file
and turns it on. If the EQ has been set using the http endpoint
then pressing SET should turn the EQ off.

http://lyra.board.ip/eq?band=1,1,1,1,1,1,1,1,0,0 -- Set the ten
band EQ to the specific values.
Sending a band=0,0,0,0,0,0,0,0,0,0 will turn the EQ off.

Signed-off-by: Volodymyr Ananiev <volodymyr.ananiev@gmail.com>